refactor: Change ContentType to enum
